pyproject.toml
[tool.poetry]
name = "redcard"
version = "0.1.0"
description = ""
authors = ["IcaroOm <icaroomdigital@gmail.com>"]
readme = "README.md"
packages = [{ include = "redcard" }]

[tool.poetry.dependencies]
python = ">=3.10"
asgiref = "3.8.1"
backports-tarfile = "1.2.0"
build = "1.2.2.post1"
cachecontrol = "0.14.2"
certifi = "2025.1.31"
cffi = "1.17.1"
charset-normalizer = "3.4.1"
cryptography = "44.0.0"
distlib = "0.3.9"
django = "5.1.5"
django-extensions = "3.2.3"
djangorestframework = "3.15.2"
drf-yasg = "1.21.8"
dulwich = "0.22.7"
fastjsonschema = "2.21.1"
filelock = "3.17.0"
idna = "3.10"
importlib-metadata = "8.6.1"
inflection = "0.5.1"
installer = "0.7.0"
jaraco-classes = "3.4.0"
jaraco-context = "6.0.1"
jaraco-functools = "4.1.0"
jeepney = "0.8.0"
keyring = "25.6.0"
more-itertools = "10.6.0"
msgpack = "1.1.0"
packaging = "24.2"
pkginfo = "1.12.0"
platformdirs = "4.3.6"
pycparser = "2.22"
pyproject-hooks = "1.2.0"
pytz = "2025.1"
pyyaml = "6.0.2"
rapidfuzz = "3.12.1"
requests = "2.32.3"
requests-toolbelt = "1.0.0"
secretstorage = "3.3.3"
shellingham = "1.5.4"
sqlparse = "0.5.3"
tomli = "2.2.1"
tomlkit = "0.13.2"
trove-classifiers = "2025.1.15.22"
typing-extensions = "4.12.2"
uritemplate = "4.1.1"
urllib3 = "2.3.0"
virtualenv = "20.29.1"
zipp = "3.21.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
